A bunch of work on the Standards interfaces
+ Create new standards. Aligned to the active course immediately.

UI work:
+ New sidebar template specifically for standards. This is definitely
a hack - the sidebar interface needs some work to be more flexible in
general.
+ Form for creating standards
+ Update forms for aligning standards to a course.
+ Create a score table for a standard-specific view. See note below.

This is definitely a workaround. Because student assessments can happen
at different times, there is no good way to header each column. I think
the best way to make it usable is to use "Attempt _n_" just to quantify
the number of assessments there have been. A hover tooltip on the score
can show the timestamp for when that assessment was submitted.