Include the standard description on the card.
This makes it easier for students to keep track of skills without
needing to remember the number schema. Also good for teachers who don't
remember the schema.

Consider moving the cards into a table. They take up a lot of vertical
space where there are a lot, but not sure how to improve that at the
moment.