Better URL path for result retrieval
The old URL path used `standards/<int> for looking up items related to a
tandard, but getting results used the result ID rather than the standard
ID the result was attached to.

This doesn't make a difference in retrieval because the standard ID
parameter isn't used in the route, but it's semantically correct when
compared to other routes.

The new route is
`/standards/<int:standard_id>/users/<int:user_id>/results/<int:result_id>`
to get a specific attempt.