Assignment results screen
Add logic and UI for details on a specific assignment. Query the
database for assessments matching the requested assignment from the
course context. Filter students out and return tuples with results
grouped by student last name.

Could probably be cleaned up, but the view works. Next steps are to add
empty spaces for students without an assessment on that given
assignment.