Proficiency work
Still trying to find a good way to represent standard progress following
#12. This will check an assignment for an existing assessment before
looking for assessment results. If there is no assessment-type
`Assignment` on a `Standard`, then the scores on their own will be used
to show progress. If an assessment is present, then it _must_ be passed
in order to be proficient.