Update chart
Show a proficiency average from the # proficient vs total enrollments
instead of calculating from the database. This needs to be sped up
somehow, so caching is put in front of the admin route to speed up
subsequent page loads.

The chart is scaled from zero to better represent relative proficiency
rates for each class.